pyproject.toml
[tool.poetry]
name = "unified-api"
version = "0.1.0"
description = ""
authors = ["ez <test@test.com>"]
readme = "README.md"
packages = [
    {include = "*", from = "rag_demo"},
    {include = "*", from = "stock_api"}
]

[tool.poetry.dependencies]
python = "^3.9"
isort = "^5.13.2"
black = "^24.4.2"
mongoengine = "0.27.0"
pyyaml = "^6.0.1"
flask-cors = "4.0.0"
python-dateutil = "^2.9.0.post0"
pymongo = "4.0"
graphene = "2.1.9"
dnspython = "2.2.1"
flask-graphql = "2.0.1"
graphene-mongo = "0.1.8"
pytest = "^8.2.2"
requests = "^2.32.3"
faiss-cpu = "1.8.0"
langchain = "^0.2.5"
openai = "^1.35.3"
langchain-community = "^0.2.5"
tiktoken = "^0.7.0"
langchain-openai = "^0.1.9"
langchain-experimental = "^0.0.61"
tabulate = "^0.9.0"
flask = "^3.0.3"
pandas = "^2.2.2"
pytest-cov = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
